refactor: separate update generation (#9788)
